Added keep-alive attribute to connections, fixed PID gain calc
TCP/IP connections can now be "kept alive" and not closed at the
end of each transaction.  This seems to improve the read/write
speeds.  ConnectionClient now has a property "keepAlive" that
can be either true or false, and this keeps the client from
closing the connection and also keeps the server from closing
the connection.

Fixed the PID gain calculation, as the calculations for Ki and Kd
were switched.