Updated PID controller, new selector for error signal, GUI changes
PID controller has been updated to a fully pipelined version with
correct handling of limits.  Gain values are now 8 bits instead
of 16 bits. This means that there is one less register for each
PID module, so the total number of PID registers is less by 2.
Software has been updated to reflect this and new limits.

There is now a selector that chooses between the two ADCs as to
which one is the error signal.  This is controlled by the topReg
bit 0.  Scan enable bit has been changed to topReg(2) instead of
topReg(16).

Scan data is now retrieved from the device and written to a binary
file and then read by the Python server.  This typically makes it
much faster.

GUI has been changed so that the two FIFOs are plotted on left and
right Y axes, and they now have their own limit controls.

When the plot settings are on "volts" one can click on the figure
and the offset will be updated to be that point.  Hopefully this
will make it easier to lock to a signal?