Changed how lock engages to avoid sudden output swings
When either PID loop is engaged, the scan module will now continue
the scan until it reaches the current "offset" value, at which
point the scan will terminate.  An "enable_o" signal is then used
to tell the PID loop that it can now engage.  This should hopefully
solve an issue I have had with real-world locking where the lock
is engaged but the system does not lock, and I think this is because
the loop engages at the same time that the voltage is swung to the
offset value, and the finite response of the system means the loop
does weird things as it fights this swing to the offset.  Now that
the scan returns to the offset value using the current scan parameters,
which are presumably set to something that gives a smooth response,
the loop should engage without drama.

Changed the GUI to add lock-in parameters.  The lock-in module is
still untested because I need a few MHz, high-voltage amplifier
to drive the EOMs.

The README has not been updated.