Simplified lock in detection, refactored MATLAB software
Lock-in detection only uses the modulation frequency, no demodulation
frequency.  Demodulation is only done with cos(\omega t), so one filter
and one multiplier has been removed.  topmod.vhd has been updated
so that there is only one "demod" option.  Software has been updated
to match.

Software refactored to use an abstract subclass for sub-modules.
Upload and fetch methods for LaserServo now use a general procedure
that doesn't need to be updated for new registers or parameters. Submodules
are compatible with this procedure as long as they are descendents of the
LaserServoSubModule class.